Extract type mapping postprocessing to an external visitor
Part of API review (#33220)
