+++ src/EFCore.Sqlite.Core/Query/Internal/SqliteQueryTranslationPostprocessor.cs
@@ -44,6 +44,15 @@ public class SqliteQueryTranslationPostprocessor : RelationalQueryTranslationPos
         return result;
     }
 
+    /// <summary>
+    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+    ///     any release. You should only use it directly in your code with extreme caution and knowing that
+    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    /// </summary>
+    protected override Expression ProcessTypeMappings(Expression expression)
+        => new SqliteTypeMappingPostprocessor(Dependencies, RelationalDependencies, RelationalQueryCompilationContext).Process(expression);
+
     private sealed class ApplyValidatingVisitor : ExpressionVisitor
     {
         protected override Expression VisitExtension(Expression extensionExpression)
